+++ Entidades/Catequista.cs
@@ -28,7 +28,7 @@ namespace Entidades
             id_Catequista = Id_Catequista;
             
         }
-        public Catequista(int Du, string Nom, string Apel, string Dir, DateTime FecNac, int Tel, char Sex, int Id_Cat) : base(Du, Nom, Apel, Dir, Tel, FecNac, Sex)
+        public Catequista(int Du, string Nom, string Apel, string Dir, DateTime FecNac, long Tel, char Sex, int Id_Cat) : base(Du, Nom, Apel, Dir, Tel, FecNac, Sex)
         {
             Dni = Du;
             Nombre = Nom;
